some updates, mainly trying to debug

+++ bsp/boards/radio.h
@@ -69,7 +69,7 @@ void                radio_setEndFrameCb(radio_capture_cbt cb);
 // reset
 void                radio_reset(void);
 // RF admin
-void                radio_setFrequency(uint8_t frequency, radio_freq_t tx_or_rx);
+void                radio_setFrequency(uint16_t frequency, radio_freq_t tx_or_rx);
 void                radio_set_modulation(RADIO_TYPE radio);
 
 //void                radio_setFrequency(uint8_t frequency);
